Add files via upload
A generic optimization criterion for optimal personalized ranking. Basically what this means is an approach that can be applied to different types of recommendation models like Matrix Factorization(MF) or k-Nearest-Neighbour (that's the “generic” part) and that solves a ranking for a set of items for a given user.The learning method is based on stochastic gradient descent with bootstrap sampling.
Whereas most other optimizers just look at if a user interacted with an item, BPR looks at the user, one item the user interacted with and one item the user did not (the unknown item). This gives us a triplet (u, i, j) of a user (u), one known item (i) and one unknown item (j).