Add files via upload
The sr method is a variant of AR, which aims to take the order of the events into 
account. Similar to the AR method, we create a sequential rule for the co-occurrence of every two items 'p' and 'q' as r(p,q) in the training data. This time, however, we consider the distance between 'p' and 'q' in the session when computing the weight of the rules. In our implementation, we use the multiplicative inverse as a weight function and set       w(p,q) = 1/x, where 'x' is the number of items that appear between 'p' and 'q' in a session.