Add files via upload
The underlying assumption on CF is that referencing items of users with similar musical taste will benefit the performance of recommender system. By adopting and
modifying this assumption, Session-based Collaborative Filtering (SSCF) finds similar sessions with an active session from given partial information and reference items in the similar sessions. Consequently, instead of making a profile for each user, we make a profile for each session as noted in section.